Fix implicit 64-to-32 bit integer narrowing
Summary:
These fire when a 64-bit integer (e.g., `size_t`, `int64_t`) is implicitly converted to a 32-bit type (e.g., `int`, `uint32_t`), which can cause data loss.

Fixes applied:
- Widened variable types to match source (e.g., `int` → `size_t`)
- Added explicit `static_cast` where narrowing is intentional and safe

Differential Revision: D97079855

+++ momentum/io/gltf/gltf_mesh_io.cpp
@@ -439,7 +439,7 @@ SkinWeights_u bindMeshToJoint(const Mesh_u& mesh, size_t jointId) {
   skinWeights->index.bottomRows(kNumVertices).setZero();
   skinWeights->weight.bottomRows(kNumVertices).setZero();
   for (auto vertId = 0; vertId < kNumVertices; vertId++) {
-    skinWeights->index(vertId, 0) = jointId;
+    skinWeights->index(vertId, 0) = static_cast<uint32_t>(jointId);
     skinWeights->weight(vertId, 0) = 1.0f;
   }
